refactor: rename requests to ddl and dml requests
